merge from source 2020-07-28

+++ src/main/java/world/bentobox/bentobox/managers/LocalesManager.java
@@ -4,16 +4,24 @@ import java.io.File;
 import java.io.FilenameFilter;
 import java.io.IOException;
 import java.io.InputStream;
+import java.io.InputStreamReader;
 import java.util.ArrayList;
 import java.util.HashMap;
+import java.util.HashSet;
 import java.util.LinkedList;
 import java.util.List;
 import java.util.Locale;
 import java.util.Map;
 import java.util.Objects;
+import java.util.Set;
 import java.util.jar.JarFile;
 
+import org.bukkit.Bukkit;
+import org.bukkit.ChatColor;
+import org.bukkit.configuration.InvalidConfigurationException;
 import org.bukkit.configuration.file.YamlConfiguration;
+import org.eclipse.jdt.annotation.NonNull;
+import org.eclipse.jdt.annotation.Nullable;
 
 import world.bentobox.bentobox.BentoBox;
 import world.bentobox.bentobox.api.addons.Addon;
@@ -31,10 +39,11 @@ public class LocalesManager {
     private Map<Locale, BentoBoxLocale> languages = new HashMap<>();
     private static final String LOCALE_FOLDER = "locales";
     private static final String BENTOBOX = "BentoBox";
+    private static final String SPACER = "*************************************************";
 
     public LocalesManager(BentoBox plugin) {
         this.plugin = plugin;
-        copyLocalesFromPluginJar(BENTOBOX);
+        copyLocalesFromPluginJar();
         loadLocalesFromFile(BENTOBOX); // Default
     }
 
@@ -44,6 +53,7 @@ public class LocalesManager {
      * @param reference a reference that can be found in a locale file
      * @return the translated String from the User's locale or from the server's locale or from the en-US locale, or null.
      */
+    @Nullable
     public String get(User user, String reference) {
         // Make sure the user is not null
         if (user != null) {
@@ -80,9 +90,11 @@ public class LocalesManager {
      * @param reference a reference that can be found in a locale file
      * @return the translated String from the server's locale or from the en-US locale, or null.
      */
+    @Nullable
     public String get(String reference) {
         // Get the translation from the server's locale
-        if (languages.get(Locale.forLanguageTag(plugin.getSettings().getDefaultLanguage())).contains(reference)) {
+        if (languages.containsKey(Locale.forLanguageTag(plugin.getSettings().getDefaultLanguage()))
+                && languages.get(Locale.forLanguageTag(plugin.getSettings().getDefaultLanguage())).contains(reference)) {
             return languages.get(Locale.forLanguageTag(plugin.getSettings().getDefaultLanguage())).get(reference);
         }
         // Get the translation from the en-US locale
@@ -93,10 +105,10 @@ public class LocalesManager {
     }
 
     /**
-     * Gets the translated String corresponding to the reference from the server's or the en-US locale file
+     * Gets the translated String corresponding to the reference from the server's or the en-US locale file.
      * or if it cannot be found anywhere, use the default text supplied.
-     * @param reference a reference that can be found in a locale file
-     * @param defaultText text to return if the reference cannot be found anywhere
+     * @param reference a reference that can be found in a locale file.
+     * @param defaultText text to return if the reference cannot be found anywhere.
      * @return the translated String from the server's locale or from the en-US locale, or default.
      */
     public String getOrDefault(String reference, String defaultText) {
@@ -105,7 +117,31 @@ public class LocalesManager {
     }
 
     /**
-     * Copies locale files from the addon jar to the file system
+     * Gets the list of prefixes from the user's locale, the server's locale and the en-US locale file.
+     * @param user the user to get the locale, not null.
+     * @return the list of prefixes from the user's locale, the server's locale and the en-US locale file.
+     * @since 1.13.0
+     */
+    public Set<String> getAvailablePrefixes(@NonNull User user) {
+        Set<String> prefixes = new HashSet<>();
+
+        // Get the player locale
+        BentoBoxLocale locale = languages.get(user.getLocale());
+        if (locale != null) {
+            prefixes.addAll(locale.getPrefixes());
+        }
+
+        // Get the prefixes from the server's locale
+        prefixes.addAll(languages.get(Locale.forLanguageTag(plugin.getSettings().getDefaultLanguage())).getPrefixes());
+
+        // Get the prefixes from the en-US locale
+        prefixes.addAll(languages.get(Locale.forLanguageTag("en-US")).getPrefixes());
+
+        return prefixes;
+    }
+
+    /**
+     * Copies locale files from the addon jar to the file system and updates current locales with the latest references
      * @param addon - addon
      */
     void copyLocalesFromAddonJar(Addon addon) {
@@ -113,38 +149,81 @@ public class LocalesManager {
             File localeDir = new File(plugin.getDataFolder(), LOCALE_FOLDER + File.separator + addon.getDescription().getName());
             if (!localeDir.exists()) {
                 localeDir.mkdirs();
-                // Obtain any locale files and save them
-                Util.listJarFiles(jar, LOCALE_FOLDER, ".yml").forEach(lf -> addon.saveResource(lf, localeDir, false, true));
             }
+            // Obtain any locale files, save them and update
+            Util.listJarFiles(jar, LOCALE_FOLDER, ".yml").forEach(lf -> {
+                File file = addon.saveResource(lf, localeDir, false, true);
+                // Update
+                if (file != null) {
+                    updateLocale(addon, file, lf);
+                }
+            });
+
         } catch (Exception e) {
             plugin.logError(e.getMessage());
+        }
+    }
+
+    private void updateLocale(Addon addon, File fileLocaleFile, String lf) {
+        try {
+            // Load the JAR locale file
+            YamlConfiguration jarLocale = addon.getYamlFromJar(lf);
+            // Load the locale file system locale file
+            YamlConfiguration fileLocale = new YamlConfiguration();
+            fileLocale.load(fileLocaleFile);
+            // Copy new keys to file
+            jarLocale.getKeys(true).stream().filter(k -> !fileLocale.contains(k, false)).forEach(k -> fileLocale.set(k, jarLocale.get(k)));
+            // Save file
+            fileLocale.save(fileLocaleFile);
+        } catch (InvalidConfigurationException e) {
+            plugin.logError("Could not update locale file '" + lf + "' due to it being malformed: " + e.getMessage());
+        } catch (Exception e) {
+            plugin.logError("Error updating locale file '" + lf + "': " + e.getMessage());
+            plugin.logStacktrace(e);
         }
     }
 
     /**
      * Copies all the locale files from the plugin jar to the filesystem.
      * Only done if the locale folder does not already exist.
-     * @param folderName - the name of the destination folder
      */
-    private void copyLocalesFromPluginJar(String folderName) {
+    private void copyLocalesFromPluginJar() {
         // Run through the files and store the locales
-        File localeDir = new File(plugin.getDataFolder(), LOCALE_FOLDER + File.separator + folderName);
+        File localeDir = new File(plugin.getDataFolder(), LOCALE_FOLDER + File.separator + LocalesManager.BENTOBOX);
         // If the folder does not exist, then make it and fill with the locale files from the jar
         // If it does exist, then new files will NOT be written!
         if (!localeDir.exists()) {
             localeDir.mkdirs();
-            FileLister lister = new FileLister(plugin);
-            try {
-                for (String name : lister.listJar(LOCALE_FOLDER)) {
-                    // We cannot use Bukkit's saveResource, because we want it to go into a specific folder, so...
-                    // Get the last part of the name
-                    int lastIndex = name.lastIndexOf('/');
-                    File targetFile = new File(localeDir, name.substring(lastIndex >= 0 ? lastIndex : 0));
-                    copyFile(name, targetFile);
+        }
+        FileLister lister = new FileLister(plugin);
+        try {
+            for (String name : lister.listJar(LOCALE_FOLDER)) {
+                // We cannot use Bukkit's saveResource, because we want it to go into a specific folder, so...
+                // Get the last part of the name
+                int lastIndex = name.lastIndexOf('/');
+                File targetFile = new File(localeDir, name.substring(Math.max(lastIndex, 0)));
+                copyFile(name, targetFile);
+                // Update the locale file if it exists already
+                try (InputStreamReader in = new InputStreamReader(plugin.getResource(name))) {
+                    YamlConfiguration jarLocale = new YamlConfiguration();
+                    jarLocale.load(in);
+
+                    YamlConfiguration fileLocale = new YamlConfiguration();
+                    fileLocale.load(targetFile);
+                    for (String k : jarLocale.getKeys(true)) {
+                        if (!fileLocale.contains(k, false)) {
+                            fileLocale.set(k, jarLocale.get(k));
+                        }
+                    }
+                    // Save it
+                    fileLocale.save(targetFile);
+                } catch (InvalidConfigurationException e) {
+                    plugin.logError("Could not update locale files from jar " + e.getMessage());
                 }
-            } catch (IOException e) {
-                plugin.logError("Could not copy locale files from jar " + e.getMessage());
+
             }
+        } catch (IOException e) {
+            plugin.logError("Could not copy locale files from jar " + e.getMessage());
         }
     }
 
@@ -154,8 +233,8 @@ public class LocalesManager {
      * @param localeFolder - locale folder location relative to the plugin's data folder
      */
     public void loadLocalesFromFile(String localeFolder) {
-        // Filter for files of length 9 and ending with .yml
-        FilenameFilter ymlFilter = (dir, name) -> name.toLowerCase(java.util.Locale.ENGLISH).endsWith(".yml") && name.length() == 9;
+        // Filter for files ending with .yml with a name whose length is >= 6 (xx.yml)
+        FilenameFilter ymlFilter = (dir, name) -> name.toLowerCase(java.util.Locale.ENGLISH).endsWith(".yml") && name.length() >= 6;
 
         // Get the folder
         File localeDir = new File(plugin.getDataFolder(), LOCALE_FOLDER + File.separator + localeFolder);
@@ -178,17 +257,16 @@ public class LocalesManager {
                     languages.put(localeObject, new BentoBoxLocale(localeObject, languageYaml));
                 }
             } catch (Exception e) {
-                BentoBox.getInstance().logError("Could not load '" + language.getName() + "' : " + e.getMessage()
+                plugin.logError("Could not load '" + language.getName() + "' : " + e.getMessage()
                 + " with the following cause '" + e.getCause() + "'." +
-                " The file has likely an invalid YML format or has been made unreadable during the process."
-                        );
+                        " The file has likely an invalid YML format or has been made unreadable during the process.");
             }
         }
     }
 
     private void copyFile(String name, File targetFile) {
         try (InputStream initialStream = plugin.getResource(name)) {
-            if (!targetFile.exists()) {
+            if (initialStream != null && !targetFile.exists()) {
                 java.nio.file.Files.copy(initialStream, targetFile.toPath());
             }
         } catch (IOException e) {
@@ -219,6 +297,16 @@ public class LocalesManager {
     }
 
     /**
+     * Returns {@code true} if this locale is available, {@code false} otherwise.
+     * @param locale the locale, not null. Consider using {@link Locale#forLanguageTag(String)} if needed.
+     * @return {@code true} if this locale is available, {@code false} otherwise.
+     * @since 1.14.0
+     */
+    public boolean isLocaleAvailable(@NonNull Locale locale) {
+        return languages.containsKey(locale);
+    }
+
+    /**
      * @return raw map of system locales to BentoBox locales
      */
     public Map<Locale, BentoBoxLocale> getLanguages() {
@@ -230,7 +318,7 @@ public class LocalesManager {
      */
     public void reloadLanguages() {
         languages.clear();
-        copyLocalesFromPluginJar(BENTOBOX);
+        copyLocalesFromPluginJar();
         loadLocalesFromFile(BENTOBOX);
         plugin.getAddonsManager().getAddons().forEach(addon -> {
             copyLocalesFromAddonJar(addon);
@@ -238,5 +326,74 @@ public class LocalesManager {
         });
     }
 
+    /**
+     * Loads all the locales available in the locale folder given.
+     * Used for loading all locales from plugin and addons.
+     *
+     * @param fix whether or not locale files with missing translations should be fixed.
+     *            Not currently supported.
+     * @since 1.5.0
+     */
+    public void analyzeLocales(boolean fix) {
+        languages.clear();
 
+        User user = User.getInstance(Bukkit.getConsoleSender());
+
+        user.sendRawMessage(ChatColor.AQUA + SPACER);
+        plugin.log(ChatColor.AQUA + "Analyzing BentoBox locale files");
+        user.sendRawMessage(ChatColor.AQUA + SPACER);
+        loadLocalesFromFile(BENTOBOX);
+        if (languages.containsKey(Locale.US)) {
+            analyze(user);
+        } else {
+            user.sendRawMessage(ChatColor.RED + "No US English in BentoBox to use for analysis!");
+        }
+        user.sendRawMessage(ChatColor.AQUA + "Analyzing Addon locale files");
+        plugin.getAddonsManager().getAddons().forEach(addon -> {
+            user.sendRawMessage(ChatColor.AQUA + SPACER);
+            user.sendRawMessage(ChatColor.AQUA + "Analyzing addon " + addon.getDescription().getName());
+            user.sendRawMessage(ChatColor.AQUA + SPACER);
+            languages.clear();
+            loadLocalesFromFile(addon.getDescription().getName());
+            if (languages.containsKey(Locale.US)) {
+                analyze(user);
+            } else {
+                user.sendRawMessage(ChatColor.RED + "No US English to use for analysis!");
+            }
+        });
+        reloadLanguages();
+    }
+
+    /**
+     *
+     * @param user - user
+     * @since 1.5.0
+     */
+    private void analyze(User user) {
+        user.sendRawMessage(ChatColor.GREEN + "The following locales are supported:");
+        languages.forEach((k,v) -> user.sendRawMessage(ChatColor.GOLD + k.toLanguageTag() + " " + k.getDisplayLanguage() + " " + k.getDisplayCountry()));
+        // Start with US English
+        YamlConfiguration usConfig = languages.get(Locale.US).getConfig();
+        // Fix config
+        YamlConfiguration fixConfig = new YamlConfiguration();
+        languages.values().stream().filter(l -> !l.toLanguageTag().equals(Locale.US.toLanguageTag())).forEach(l -> {
+            user.sendRawMessage(ChatColor.GREEN + SPACER);
+            user.sendRawMessage(ChatColor.GREEN + "Analyzing locale file " + l.toLanguageTag() + ":");
+            YamlConfiguration c = l.getConfig();
+            boolean complete = true;
+            for (String path : usConfig.getKeys(true)) {
+                if (!c.contains(path, true)) {
+                    complete = false;
+                    fixConfig.set(path, user.getTranslationOrNothing(path).replace('§', '&'));
+                }
+            }
+            if (complete) {
+                user.sendRawMessage(ChatColor.GREEN + "Language file covers all strings.");
+            } else {
+                user.sendRawMessage(ChatColor.RED + "The following YAML is missing. Please translate it:");
+                plugin.log("\n" + fixConfig.saveToString());
+            }
+
+        });
+    }
 }
